requirements.txt
certifi==2019.3.9
chardet==3.0.4
Django==2.1.7
django-ckeditor==5.6.1
django-js-asset==1.2.1
idna==2.8
mysqlclient==1.4.2
Pillow==5.4.1
pytz==2018.9
requests==2.21.0
urllib3==1.24.1
